Revert "Cleanup branch merge"
