+++ algorithms/ball_tracking.py
@@ -1,0 +1,137 @@
+# Ball tracking example from 
+# http://www.pyimagesearch.com/2015/09/14/ball-tracking-with-opencv/
+
+# import the necessary packages
+from collections import deque
+import numpy as np
+import argparse
+#import imutils
+import cv2
+
+import signal
+import sys
+
+# Allow safe exit on ctrl-c
+exit_flag = False
+def signal_handler(signal, frame):
+    global exit_flag
+    exit_flag = True
+    print('Exiting')
+signal.signal(signal.SIGINT, signal_handler)
+
+# construct the argument parse and parse the arguments
+ap = argparse.ArgumentParser()
+ap.add_argument("-v", "--video",
+    help="path to the (optional) video file")
+ap.add_argument("-b", "--buffer", type=int, default=64,
+    help="max buffer size")
+args = vars(ap.parse_args())
+
+# define the lower and upper boundaries of the "green"
+# ball in the HSV color space, then initialize the
+# list of tracked points
+greenLower = np.array(cv2.cv.Scalar(29, 86, 6))
+greenUpper = np.array(cv2.cv.Scalar(64, 255, 255))
+minRadius  = 20
+pts = deque(maxlen=args["buffer"])
+
+# if a video path was not supplied, grab the reference
+# to the webcam
+if not args.get("video", False):
+    camera = cv2.VideoCapture(0)
+    camera.set(cv2.cv.CV_CAP_PROP_FRAME_WIDTH, 640)
+    camera.set(cv2.cv.CV_CAP_PROP_FRAME_HEIGHT, 480)
+    
+# otherwise, grab a reference to the video file
+else:
+    camera = cv2.VideoCapture(args["video"])
+
+# grab an output video file
+fourcc = cv2.cv.CV_FOURCC(*'XVID')
+out = cv2.VideoWriter('video.avi', fourcc, 20, (640, 480))
+assert(out.isOpened())
+
+# keep looping
+while not exit_flag:
+    # grab the current frame
+    (grabbed, frame) = camera.read()
+    if not grabbed:
+        print "Frame capture failed"
+    frame = cv2.flip(frame, 1)
+    #cv2.imshow("Frame", frame) 
+
+    # if we are viewing a video and we did not grab a frame,
+    # then we have reached the end of the video
+    if args.get("video") and not grabbed:
+        break
+
+    # resize the frame, blur it, and convert it to the HSV
+    # color space
+    # frame = imutils.resize(frame, width=600)
+    # blurred = cv2.GaussianBlur(frame, (11, 11), 0)
+    hsv = cv2.cvtColor(frame, cv2.COLOR_BGR2HSV)
+
+    # construct a mask for the color "green", then perform
+    # a series of dilations and erosions to remove any small
+    # blobs left in the mask
+    
+    mask = cv2.inRange(hsv, greenLower, greenUpper)
+    mask = cv2.erode(mask, None, iterations=2)
+    mask = cv2.dilate(mask, None, iterations=2)
+
+    # find contours in the mask and initialize the current
+    # (x, y) center of the ball
+    cnts = [] # TEMP: Disable contour detection
+    cnts = cv2.findContours(mask.copy(), cv2.RETR_EXTERNAL,
+        cv2.CHAIN_APPROX_SIMPLE)[-2]
+    center = None
+
+    # only proceed if at least one contour was found
+    if len(cnts) > 0:
+        # find the largest contour in the mask, then use
+        # it to compute the minimum enclosing circle and
+        # centroid
+        c = max(cnts, key=cv2.contourArea)
+        ((x, y), radius) = cv2.minEnclosingCircle(c)
+        M = cv2.moments(c)
+        center = (int(M["m10"] / M["m00"]), int(M["m01"] / M["m00"]))
+
+        # only proceed if the radius meets a minimum size
+        if radius > minRadius:
+            # draw the circle and centroid on the frame,
+            # then update the list of tracked points
+            cv2.circle(frame, (int(x), int(y)), int(radius),
+                (0, 255, 255), 2)
+            cv2.circle(frame, center, 5, (0, 0, 255), -1)
+            print "Ball Detected at (%d,%d), radius %d" % (x,y,radius) 
+            
+    # update the points queue
+    pts.appendleft(center)
+
+    # loop over the set of tracked points
+    for i in xrange(1, len(pts)):
+        # if either of the tracked points are None, ignore
+        # them
+        if pts[i - 1] is None or pts[i] is None:
+            continue
+
+        # otherwise, compute the thickness of the line and
+        # draw the connecting lines
+        thickness = int(np.sqrt(args["buffer"] / float(i + 1)) * 2.5)
+        cv2.line(frame, pts[i - 1], pts[i], (0, 0, 255), thickness)
+
+    # show the frame to our screen
+    #cv2.imshow("Frame", frame)
+    # save the frame to a file
+    out.write(frame)
+
+    key = cv2.waitKey(1) & 0xFF
+
+    # if the 'q' key is pressed, stop the loop
+    if key == ord("q"):
+        break
+
+# cleanup the camera and close any open windows
+camera.release()
+out.release()
+cv2.destroyAllWindows()
