+++ drivers/hmc5883l.py
@@ -1,0 +1,122 @@
+#!/usr/bin/env python
+# vim: set fileencoding=UTF-8 :
+
+# Author: rm-hull
+# Source: https://github.com/rm-hull/hmc5883l
+
+# HMC5888L Magnetometer (Digital Compass) wrapper class
+# Based on https://bitbucket.org/thinkbowl/i2clibraries/src/14683feb0f96,
+# but uses smbus rather than quick2wire and sets some different init
+# params.
+
+import smbus2 as smbus
+import math
+import time
+import sys
+
+class hmc5883l:
+    # Calibration
+    # TODO: Let the user set this
+    # Source for these values: followed the procedure at 
+    # http://blog.bitify.co.uk/2013/11/connecting-and-calibrating-hmc5883l.html
+    __x_offset = -161.0
+    __y_offset = 164.0
+    __x_scale  = 226.0
+    __y_scale  = 215.0
+    
+    __scales = {
+        0.88: [0, 0.73],
+        1.30: [1, 0.92],
+        1.90: [2, 1.22],
+        2.50: [3, 1.52],
+        4.00: [4, 2.27],
+        4.70: [5, 2.56],
+        5.60: [6, 3.03],
+        8.10: [7, 4.35],
+    }
+
+    def __init__(self, port=1, address=0x1E, gauss=1.3, declination=(0,0)):
+        self.bus = smbus.SMBus(port)
+        self.address = address
+
+        (degrees, minutes) = declination
+        self.__declDegrees = degrees
+        self.__declMinutes = minutes
+        self.__declination = (degrees + minutes / 60) * math.pi / 180
+
+        (reg, self.__scale) = self.__scales[gauss]
+        self.bus.write_byte_data(self.address, 0x00, 0x70) # 8 Average, 15 Hz, normal measurement
+        self.bus.write_byte_data(self.address, 0x01, reg << 5) # Scale
+        self.bus.write_byte_data(self.address, 0x02, 0x00) # Continuous measurement
+
+    def declination(self):
+        return (self.__declDegrees, self.__declMinutes)
+
+    def twos_complement(self, val, len):
+        # Convert twos compliment to integer
+        if (val & (1 << len - 1)):
+            val = val - (1<<len)
+        return val
+
+    def __convert(self, data, offset):
+        val = self.twos_complement(data[offset] << 8 | data[offset+1], 16)
+        if val == -4096: return None
+        return round(val * self.__scale, 4)
+
+    def axes(self):
+        data = self.bus.read_i2c_block_data(self.address, 0x00)
+        #print map(hex, data)
+        x = self.__convert(data, 3)
+        y = self.__convert(data, 7)
+        z = self.__convert(data, 5)
+        return (x,y,z)
+
+    def heading(self):
+        (x, y, z) = self.axes()
+        x = (x + self.__x_offset) / self.__x_scale
+        y = (y + self.__y_offset) / self.__y_scale
+        headingRad = math.atan2(y, x)
+        headingRad += self.__declination
+
+        # Correct for reversed heading
+        if headingRad < 0:
+            headingRad += 2 * math.pi
+
+        # Check for wrap and compensate
+        elif headingRad > 2 * math.pi:
+            headingRad -= 2 * math.pi
+
+        # Convert to degrees from radians
+        headingDeg = math.degrees(headingRad)
+        return headingDeg
+
+    def degrees(self, headingDeg):
+        degrees = math.floor(headingDeg)
+        minutes = round((headingDeg - degrees) * 60)
+        return (degrees, minutes)
+
+    def __str__(self):
+        (x, y, z) = self.axes()
+        return "Axis X: " + str(x) + "\n" \
+               "Axis Y: " + str(y) + "\n" \
+               "Axis Z: " + str(z) + "\n" \
+               "Declination: " + self.degrees(self.declination()) + "\n" \
+               "Heading: " + self.degrees(self.heading()) + "\n"
+
+if __name__ == "__main__":
+    compass = hmc5883l(gauss = 1.3, declination = (-175,0))
+    
+    (raw_x, raw_y, raw_z) = compass.axes()
+    (max_x, max_y) = (raw_x, raw_y)
+    (min_x, min_y) = (raw_x, raw_y)
+    
+    while True:
+        (raw_x, raw_y, raw_z) = compass.axes()
+        (max_x, max_y) = (max(raw_x, max_x), max(raw_y, max_y))
+        (min_x, min_y) = (min(raw_x, min_x), min(raw_y, min_y))
+        sys.stdout.write("\rHeading: " + str(compass.heading()) + "     ")
+        sys.stdout.flush()
+        time.sleep(0.5)
+    print("Raw:", (raw_x, raw_y))
+    print("Calibration min(x,y), max(x,y): ", (min_x, min_y), (max_x, max_y))
+
