+++ mag_calibration.json
@@ -1,1 +1,1 @@
-{"max_x": 0.22519727168196943, "due_north_offset": 84.78201257971203, "min_x": 0.0464105174111391, "min_y": 0.34978699260905266, "max_y": 0.43012765082622084}
+{"due_north_offset": 274.1936718756784, "min_x": -0.1525055079586358, "min_y": -0.23744222753576066, "max_x": 0.28960442823975097, "max_y": 0.21971799862945748}
