+++ pibakery_recipe.xml
@@ -1,0 +1,1 @@
+<xml xmlns="http://www.w3.org/1999/xhtml"><block type="onfirstboot" id="onfirstboot" x="77" y="31"><next><block type="uartconsole" id="HiN4G^WF_g8@Tt!@Zw#*"><field name="1">Disable</field><next><block type="runcommand" id="B~YbBXepmP),THy2ntDO"><field name="1">sed -i '/enable_uart/s/= .*/=1/' /boot/config.txt</field><field name="2">root</field><next><block type="setspi" id="V?ZG_;HN#J]LD-,PbR@1"><field name="1">Enable</field><next><block type="seti2c" id="i;t%^u*McrGc{02`{0!F"><field name="1">Enable</field><next><block type="sethostname" id="Q?jp{)t1.hAd5OeoQTn5"><field name="1">autonomyPi</field><next><block type="wifisetup" id="dM~orGp8Hd~~ju`i8V2h"><field name="1">MST-PSK-N</field><field name="2">JoeMiner</field><field name="3">WPA/WPA2</field><next><block type="runcommand" id="d2Kx{rr,@BsBJxQGKx:Z"><field name="1">sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" &gt; /etc/apt/sources.list.d/ros-latest.list'</field><field name="2">root</field><next><block type="runcommand" id="Evw4pK^Ahj51=9{Khj[e"><field name="1">apt-key adv --keyserver hkp://ha.pool.sks-keyservers.net:80 --recv-key 0xB01FA116</field><field name="2">root</field><next><block type="runcommand" id="it#vG;CObl-OH0`uZ42?"><field name="1">apt-get update</field><field name="2">root</field><next><block type="packageinstall" id="S*QRd{8AfCOL:%#4dF|E"><field name="1">ros-kinetic-full</field><next><block type="runcommand" id="_1Rv]Q[Nytw{NOgk~Q?l"><field name="1">rosdep init</field><field name="2">root</field><next><block type="runcommand" id="{yLWnK)^btc=XFI})h4u"><field name="1">rosdep update</field><field name="2">pi</field><next><block type="runcommand" id="wcs?4/|M@D[Q06OP[dra"><field name="1">echo "source /opt/ros/kinetic/setup.bash" &gt;&gt; ~/.bashrc</field><field name="2">pi</field><next><block type="runcommand" id="hB%VBtWancuqJdDEF{yG"><field name="1">git clone https://github.com/MST-MRDT/AutonomyBoard-Software.git</field><field name="2">pi</field><next><block type="runcommand" id="|DC0id!h`]ltuYpn|.o7"><field name="1">pip install Adafruit_PCA9685 pynmea2</field><field name="2">root</field><next><block type="packageinstall" id="woVR@)YiAN37M+juWpb{"><field name="1">python-opencv</field><next><block type="reboot" id="cRL)jk*Gv@Q9!%p*=Ram"></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></xml>
