+++ requirements.txt
@@ -1,0 +1,54 @@
+Adafruit-GPIO==1.0.1
+Adafruit-PCA9685==1.0.1
+Adafruit-PureIO==0.2.1
+Flask==0.10.1
+Jinja2==2.7.3
+MarkupSafe==0.23
+Pillow==2.6.1
+RPi.GPIO==0.6.2
+RTIMULib==7.2.1
+Werkzeug==0.9.6
+argparse==1.2.1
+astroid==1.2.1
+blinker==1.3
+chardet==2.3.0
+colorama==0.3.2
+gpiozero==1.3.1
+html5lib==0.999
+itsdangerous==0.24
+logilab-common==0.62.0
+lxkeymap==0.1
+matplotlib==1.4.2
+mcpi==0.1.1
+mock==1.0.1
+ndg-httpsclient==0.3.2
+nose==1.3.4
+numpy==1.8.2
+picamera==1.12
+picraft==0.6
+pifacecommon==4.2.1
+pifacedigitalio==3.1.0
+pigpio==1.30
+pyOpenSSL==0.13.1
+pyasn1==0.1.7
+pygame==1.9.2a0
+pygobject==3.14.0
+pyinotify==0.9.4
+pylint==1.3.1
+pynmea2==1.7.1
+pyparsing==2.0.3
+pyserial==2.6
+python-apt==0.9.3.12
+python-dateutil==2.2
+pytz==2012c
+requests==2.4.3
+sense-emu==1.0
+sense-hat==2.2.0
+six==1.8.0
+smbus==1.1
+spidev==3.0
+spyder==2.3.1
+twython==3.1.2
+urllib3==1.9.1
+wheel==0.24.0
+wsgiref==0.1.2
